Several AD fixes
Fixed a few issues with lamda_init_2 integration. Fixed a bug with alg_depth = 2 that didn't allow the training to update. Applied Jofre's fix (tensor.modify) for the update bug that was also present in alg_depth 1. added annotations around using dask in the example and details in the notebooks.